+++ CHANGES.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added `Lua.set_global/3`, `Lua.get_global/2`.
 - Added `Lua.set_package_path/2`.
 - Added `Lua.require!/2`.
+- Added `Lua.encode/1` and `Lua.decode/1`.
 
 ## [0.3.0] - 2016-08-08
 ### Added